modules/SampleModule/requirements.txt
azure-iot-device
paho-mqtt